fix: re-assign local audio management to global
The following changes were made to resolve broken audio state.
- Assign local audio management to global manager(Zustand).
This resolves the inconsistency with the audio output when the music
player mounts.
The fix ensures that, there's a global management of the "new audio"
setter managed by Zustand and will be available to all components that
will need to use the "audio setter".